Simplified the sample application code to only include the minimum set of components required to illustrate how to create a NestJS app that leverages the node-abstract-repository library.
