+++ examples/nestjs-mongoose-book-manager/package.json
@@ -29,7 +29,7 @@
     "@nestjs/platform-express": "^9.0.0",
     "class-transformer": "^0.5.1",
     "mongoose": "^7.3.2",
-    "node-abstract-repository": "link:../../",
+    "node-abstract-repository": "^2.2.0",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.2.0",
     "typescript-optional": "^3.0.0-alpha.3"
