+++ examples/nestjs-mongoose-book-manager/src/book.repository.ts
