+++ examples/nestjs-mongoose-book-manager/src/book.schemas.ts
