+++ examples/nestjs-mongoose-book-manager/src/book.ts
