--- examples/nestjs-mongoose-book-manager/src/book/book.module.ts
@@ -1,16 +1,0 @@
-import { Module } from '@nestjs/common';
-import { BookService } from './book.service';
-import { BookController } from './book.controller';
-import { MongooseBookRepository } from './book.repository';
-
-@Module({
-  providers: [
-    {
-      provide: 'BOOK_REPOSITORY',
-      useClass: MongooseBookRepository,
-    },
-    BookService,
-  ],
-  controllers: [BookController],
-})
-export class BookModule {}
