--- examples/nestjs-mongoose-book-manager/src/book/book.service.ts
@@ -1,31 +1,0 @@
-import { Inject, Injectable } from '@nestjs/common';
-import { Book } from './book';
-import { Repository } from 'node-abstract-repository';
-
-export type PartialBook = { id: string } & Partial<Book>;
-export type PersistentBook = Book | PartialBook;
-
-@Injectable()
-export class BookService {
-  constructor(
-    @Inject('BOOK_REPOSITORY')
-    private readonly bookRepository: Repository<Book>,
-  ) {}
-
-  async save(book: PersistentBook): Promise<Book> {
-    if (book) {
-      try {
-        return await this.bookRepository.save(book);
-      } catch (error) {}
-    }
-    return null as unknown as Book;
-  }
-
-  async findAll(): Promise<Book[]> {
-    return this.bookRepository.findAll();
-  }
-
-  async deleteById(id: string): Promise<boolean> {
-    return this.bookRepository.deleteById(id);
-  }
-}
